+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+### Fixed
+- Add scorer traces to correct parent sample when running with multiple epochs
+
 ## [v0.1.6](https://pypi.org/project/inspect-wandb/0.1.6/) (23 September 2025)
 
 ### Added